test: stability fixes and pg stat collection support
