+++ src/test/java/io/supertokens/test/ConfigTest2_6.java
@@ -198,23 +198,35 @@ public class ConfigTest2_6 {
 
     @Test
     public void testThatMissingConfigFileThrowsError() throws Exception {
+        // Skip this test in parallel execution as it modifies the global config.yaml file
+        String workerId = System.getProperty("org.gradle.test.worker", "");
+        Assume.assumeTrue("Skipping in parallel test execution - modifies global config.yaml",
+                workerId.isEmpty());
+
         String configFile = new File(new File(".").getAbsoluteFile().getParentFile().getParentFile().getPath(), "config.yaml").getAbsolutePath();
         String[] args = {"../", "configFile=" + configFile};
 
-        ProcessBuilder pb = new ProcessBuilder("rm", "config.yaml");
-        pb.directory(new File(args[0]));
-        Process process1 = pb.start();
-        process1.waitFor();
+        try {
+            ProcessBuilder pb = new ProcessBuilder("rm", "config.yaml");
+            pb.directory(new File(args[0]));
+            Process process1 = pb.start();
+            process1.waitFor();
 
-        TestingProcess process = TestingProcessManager.startIsolatedProcess(args);
-        ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
-        assertNotNull(e);
-        assertEquals(e.exception.getMessage(),
-                configFile + " (No such file or directory)");
+            TestingProcess process = TestingProcessManager.startIsolatedProcess(args);
+            ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
+            assertNotNull(e);
+            assertEquals(e.exception.getMessage(),
+                    configFile + " (No such file or directory)");
 
-        process.kill();
-        assertNotNull(process.checkOrWaitForEvent(PROCESS_STATE.STOPPED));
-
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(PROCESS_STATE.STOPPED));
+        } finally {
+            // Restore config.yaml from temp
+            ProcessBuilder pb = new ProcessBuilder("cp", "temp/config.yaml", "config.yaml");
+            pb.directory(new File(args[0]));
+            Process process1 = pb.start();
+            process1.waitFor();
+        }
     }
 
     @Test
