+++ src/test/java/io/supertokens/test/TestHelloAPIRateLimiting.java
@@ -153,7 +153,8 @@ public class TestHelloAPIRateLimiting {
     public void testThatTheHelloAPIisRateLimited() throws Exception {
         String[] args = {"../"};
 
-        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args);
+        // Use startProcess=false to avoid race condition with feature flag setup
+        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args, false);
         FeatureFlagTestContent.getInstance(process.getProcess())
                 .setKeyValue(FeatureFlagTestContent.ENABLED_FEATURES, new EE_FEATURES[]{EE_FEATURES.MULTI_TENANCY});
         process.startProcess();
@@ -185,7 +186,8 @@ public class TestHelloAPIRateLimiting {
     public void testThatTheHelloAPIisRateLimitedPerApp() throws Exception {
         String[] args = {"../"};
 
-        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args);
+        // Use startProcess=false to avoid race condition with feature flag setup
+        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args, false);
         FeatureFlagTestContent.getInstance(process.getProcess())
                 .setKeyValue(FeatureFlagTestContent.ENABLED_FEATURES, new EE_FEATURES[]{EE_FEATURES.MULTI_TENANCY});
         process.startProcess();
@@ -229,7 +231,8 @@ public class TestHelloAPIRateLimiting {
     public void testThatTheHelloAPIisRateLimited2() throws Exception {
         String[] args = {"../"};
 
-        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args);
+        // Use startProcess=false to avoid race condition with feature flag setup
+        TestingProcessManager.TestingProcess process = TestingProcessManager.startIsolatedProcess(args, false);
         FeatureFlagTestContent.getInstance(process.getProcess())
                 .setKeyValue(FeatureFlagTestContent.ENABLED_FEATURES, new EE_FEATURES[]{EE_FEATURES.MULTI_TENANCY});
         process.startProcess();
